code/HearoMQTT/requirements.txt
paho-mqtt<2